Add language semantics
Mostly Russian, a smattering of French and Latin. This also adds a few other missing semantic attributes and accents. Having done that, the sentece in the introduction about the removal of foreign word italicisation and accents no longer makes any sense and has been removed.

Russian words in this text that have been deemed common enough in English to remain unitalicised:
- rouble
- kopeck
- samovar
- mazurka
- hussar